requirements.txt
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
pyspark==3.5.3
scikit-learn>=1.2.0
scikit-surprise>=1.1.3
xgboost>=1.7.0
lightfm>=1.17
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

pytest>=7.3.0
pyarrow>=12.0.0
fastparquet>=2023.4.0
flask>=2.3.0
